Update patch set 8
Patch Set 8: Code-Review+2

There are still some small things I do not like, such as "CONTROLLERID" (could be "MEMBER_ID"), "Controller is UP" (could be "ODL is UP") or "VAR" in the commit message (could be "variable").

But it is not worth delaying this change any further.

Patch-set: 8
Label: Code-Review=+2