Update patch set 2
Patch Set 2:

> Dont you need sudo for this command?

It is not a standalone command, it is bash sub-command. Sudo needs standalone command to execute and of course if I start with "sudo bash" it would spawn new bash and not affect the current bash.

It is possible that something more low-level is needed.

Patch-set: 2
Label: Verified=0