Update patch set 6
Patch Set 6: Code-Review+1

This is cleaned-up version of patch set 5 which was able to go on with the AAA test in Sandbox (although test cases failed to establish connection, possibly due to using port 8282). It also shows that pip is upgraded from 1.4.1 to 7.1.2:
https://jenkins.opendaylight.org/sandbox/job/aaa-csit-1node-idmlite-only-beryllium/4/console

Run #5 tests patch set 6. There are some warnings printed, which I leave as TODO.

Patch-set: 6
Label: Code-Review=+1