Update patch set 2
Patch Set 2:

> remove openflowplugin plugin from the list of compatible features and that does not seem right

Well, *-li features may be not the default choice, but it is still an official deliverable, we are system-testing it, and some users may use it and expect it to work together with netconf, pcep and similar independent plugins.

Maybe we could rename odl-integration-compatible-with-all to odl-integration-compatible-with-all-default so that openflowplugin remains there (or use both lists), but we should not hide the fact that there are basically two mutually incompatible openflow plugins.

I would actually prefer moving openflow to incompatible list; so that projects like l2switch may start discussing whether it is possible for their single feature to interoperate with both openflowplugin implementations, which would return it back to odl-integration-compatible-with-all.

By the way, I am no opposed to reducing number of install=all jobs, I just do not like the motivation as stated in commit message.

Patch-set: 2