Update patch set 1
Patch Set 1:

This would be better as a Robot suite.
But if the point is to abort quickly, it would not work in current setup with single permissive pybot invocation.
I have ideas how to rework it for several pybot invocations, but I do not see myself having time for that in foreseeable future.

So, this change is good enough for Beryllium.

Patch-set: 1