Update patch set 2
Patch Set 2: -Code-Review

| Well, *-li features may be not the default choice, but it is still an official deliverable, we are system-testing it, and some users may use it and expect it to work together with netconf, pcep and similar independent plugins.

How the -all- test works today, I do not see an easy way to run netconf or pcep tests for both helium and li plugins. Moreover if we remove openflow default plugin from compatible list, we will never test these features in presence of openflow.

| Maybe we could rename odl-integration-compatible-with-all to odl-integration-compatible-with-all-default so that openflowplugin remains there (or use both lists), but we should not hide the fact that there are basically two mutually incompatible openflow plugins.

So basically you are proposing changing the current feature classification and duplicating the number of -all- tests so that one set will include the old plugin and the other set will include the new plugin. It is feasible but it will require a lot of work including changing the existing process for projects to add their features to integration.

| I would actually prefer moving openflow to incompatible list; so that projects like l2switch may start discussing whether it is possible for their single feature to interoperate with both openflowplugin implementations, which would return it back to odl-integration-compatible-with-all.

I may be wrong but my understanding is projects would never have to modify their features for the new plugin, the plugin itself will swap to the new implementation. I will probably confirm this with the openflowplugin people. The only reason for moving openflow and all features depending on it to not compatible will be to test the new Li plugin with other features at the price of not testing other features with openflow.

| By the way, I am no opposed to reducing number of install=all jobs, I just do not like the motivation as stated in commit message.

The fact is that these 2 jobs are bringing both plugins so it does not makes sense to run this test today. I can change the commit message to something more on this line or I can abandon this patch and continue having 2 meaningless tests until we decide what to do with the openflowplugin features.

Patch-set: 2
Reviewer: Gerrit User 797 <797@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0