Update patch set 2
Patch Set 2:

DEVSTACK_GATE_TIMEOUT is a flag inside the devstackgate shell scripts so it is independent of the BUILD_TIMEOUT at 180m. It defaults to 60 but I think somewhere it was set to 120. You export the value to set it for the scripts. In the console logs you can see "Job timeout set to: 110 minutes" which is the value remaining.

It is set with REMAINING_TIME=$((DEVSTACK_GATE_TIMEOUT - elapsed - 5))
in the scripts. Notice the 5 at the end which looked a lot like the nice round 5's you were seeing. But even this value is coming up longer than the 50-105 minutes where it is failing.

It also uses ntp to get the time so maybe something is wrong in our ntp setup?

But since they already timeout on this 5 multiple maybe there are other places that have a similar timeout.

Patch-set: 2
Reviewer: Gerrit User 773 <773@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0