Update patch set 1
Patch Set 1:

> general purpose suite teardown

I do not like actual tests in teardown, because: "If the suite teardown fails, all tests in the suite are marked failed afterwards in reports and logs." [0]

In other words, it hides passrate information and provides no speedup.
But such general keyword can be used in last ordinary testcase.
Also, it can be used in Setup to skip all tests.

[0] http://robotframework.org/robotframework/latest/RobotFrameworkUserGuide.html#setups-and-teardowns

Patch-set: 1