Update patch set 1
Patch Set 1: Code-Review+2 Verified+1

Ok, so this looks good and I've updated the image in both sandbox and releng. As an FYI for some reason the final udpreplay binary is 0 bytes in size in releng though I put some checks around it during my spin up to see if there was a problem. The as built binary is good. So, for releng I added an extra forced recopy of it to the unmanaged portion of the spin-up scripts, hopefully that gets it working correctly there.

Patch-set: 1
Reviewer: Gerrit User 1 <1@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=+1
Label: Code-Review=+2