Configure clustering on a list of controllers.
Adding a new script called include-raw-integration-configure-clustering-list.sh,
which is a list version of include-raw-integration-configure-clustering.sh.
The older script worked for exactly 3 controllers. The new script takes a list of
controllers using an environment variable called ODL_SYSTEM_IP_LIST. This would
have a comma-separated list of controllers. This allows any arbitrary number
of controllers to be configured as part of a cluster.

Change-Id: Iaebf8fdedc0998974733c896de97d51bbaf3e452
Signed-off-by: Shaleen Saxena <ssaxena@brocade.com>