Vagrant and start-up script updates for GBP image
This change is a working setup for the image that is proposed by GBP for
testing purposes.

NOTE: Due to snapshots refusing to connect to the network after OVS is
installed but not having an issue once the system is up and OVS is
installed, the final OVS installation / statup is left as a step to
happen in the jenkins-script. That change is part of this submission.

Change-Id: I0cd2861d6e8cfe50538a9064eb30ca44b9d70153
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>