Update patch set 2
Patch Set 2: Code-Review+2

> continue having 2 meaningless tests

Oh, when you put it this way, the commit message is not strong enough reason for continued waste of CI resources.

As a refactor of -all- jobs is planned anyway, we can discuss details of OPF there. I have edited Trello card [0] as a reminder, so now I see no reason to postpone merging.

[0] https://trello.com/c/wJJYalmk/229-create-a-all-job-that-runs-multiple-project-suites-back-to-back-no-controller-re-install-actually-two-such-jobs-one-for-each-ope

Patch-set: 2
Label: Code-Review=+2