Update patch set 13
Patch Set 13: Code-Review+1

Cool, so I think the only thing missing is netopeer but as I commented before we are going to deprecate this tool and instead use the netconf tool created by netconf developers. We could ask tomorrow in the integration call ETA for the new netconf test and we can have the curent netconf test running on the old centos vm until then.

Patch-set: 13