Update patch set 13
Patch Set 13:

> Cool, so I think the only thing missing is netopeer but as I
 > commented before we are going to deprecate this tool and instead
 > use the netconf tool created by netconf developers. We could ask
 > tomorrow in the integration call ETA for the new netconf test and
 > we can have the curent netconf test running on the old centos vm
 > until then.

So the consensus is that we can keep using the existing VM for the one 
CSIT job that needs netopeer tool, and migrate all the others to use
this ubuntu mininet VM.

Carol mentioned she may be able to get that netconf test ported
to the testtool at some point, and after that we can be done with
the existing mininet VM if that's the goal.

In the end, I think this is good to go now.

Patch-set: 13