Update patch set 2
Patch Set 2:

Radovan had said that the other suites also have similar behavior: on Releng they pass, but on Sandbox if there are multiple performance jobs run at once, they timeout.
The fix should be done in integration/test, as noted in FIXME. After that, this job definition may change (either just delete FIXME or also include the new option).

For the option value, we would like to see the performance of the new suite on Releng, that is why I am merging this.

Patch-set: 2