Update patch set 1
Patch Set 1: Code-Review+1

Based on Jenkins daily job results this change defines the value of the test suite robot option TEST_DURATION_INDEX which has impact on (multiplies) the originally calculated timeout for filling the topology in the simplest - one prefix - scenario. This option is introduced by the Change 29090 (to be merged after the 29090 please).

Patch-set: 1
Label: Code-Review=+1