Support any number of controller and mininet instances
Following Shaleen works, I have updated cluster scripts to support
any number of controller and mininet instances.

Cluster scripts will use new env variable CONTROLLERCOUNT and
MININETCOUNT to loop the instances.

This is a first patch and we will need further patches to allow
users to select the number of controller instances as well as the
cluster templates location.

Change-Id: I5bb32024841fde2afc9c5c7183a83f0cc078d8f8
Signed-off-by: Luis Gomez <ecelgp@gmail.com>