Update patch set 2
Patch Set 2:

(1 comment)

Thanks for looking at this Venkat.   This job as a whole has several problems, but the
biggest is that it hangs while running tempest tests.  you can see that happen in the 
sandbox and releng.  I'm looking for workarounds to that (see fyi below).  I have not
even begun to debug any specific tempest failures, but it sounds like this change to
not create initial networks could help.



Also, just fyi, but I'm currently refactoring this job and will be posting a larger patchset for that (soon hopefully).  You can look at it now in the sandbox, although it's not cleaned up yet since I'm still debugging:

https://jenkins.opendaylight.org/sandbox/job/ovsdb-daily-openstack-beryllium/

Patch-set: 2