Update patch set 13
Patch Set 13: Code-Review+1

The new variables are defined in Change 29205, but this and that change can be merged in any order.

Sandbox example: https://jenkins.opendaylight.org/sandbox/job/bgpcep-csit-1node-periodic-bgp-ingest-only-stable-lithium/15

The last suite fails because SSH connections get dropped. I believe that is an environment bug, but I am not sure how to examine it further.
I think it would be useful to have it running in this state as a reminder that something needs to be examined and fixed.

Patch-set: 13