Update patch set 1
Patch Set 1:

This is working in a hacked up sandbox job [0].  I wanted to get this shell script up for review now, and I'll continue to work on the jjb work that is still needed.  I have a feeling there will be some things to address here, so the sooner the better.

Patch-set: 1
Label: Verified=0