Update patch set 1
Patch Set 1:

Is the test suite principialy incompatible with Lithium, or does it just need some fixing?
If the suite (or ODL) would be fixed before final Lithium SR, we should keep the job as a reminder of a debt.
But if fixing has low priority, we can remove the job anyway to free VMs.

Patch-set: 1