Update patch set 2
Patch Set 2:

I tested this against ttp as a test project [1]. Seems to take just over 10 minutes to run. I can't seem to run validate in parallel though as it keeps throwing an error when I try which is too bad because it would shave off a lot of time. This patch adds a new CFG configuration called "AUTORELEASE" and when set to true it will add a project-validate-autorelease job to each project's auto templates.

[1] https://jenkins.opendaylight.org/releng/job/ttp-validate-autorelease-beryllium/10/

Patch-set: 2
Label: Verified=0