Update patch set 1
Patch Set 1:

Hmm I'll take a look at the sandbox when I get a chance. What's the benefit of running pip freeze though?

I usually use it as a starting point for creating requirements.txt but looks like you just want to know the versions of things we have installed when this job is run?

Patch-set: 1