Update patch set 1
Patch Set 1:

We need both, the old version for backwards compatibility and the lastest version for new tests. Unless we keep the c6 vms for backwards compatibility. I guess it is up to you to decide which VM you would like to maintain for old test: c6 mininet or ubuntu mininet with old OVS.

Patch-set: 1