Update patch set 1
Patch Set 1:

Thanks Vratko for bringing this up. I'll have to update the job so that it properly clones the new integration sub-projects.

I'll consider improving this to scan multiple streams but initially this was meant to be run against only the master branch, but running against supported streams might be useful data point too.

Patch-set: 1