Update patch set 2
Patch Set 2:

And why we need this? If we use a different branch in the distribution it will not pull the projects artifacts created by this job, right? I saw some comment from Vratko regarding the ability of passing a PATCHREFSPEC to the integration-distribution-test-{stream} job, but that is not this change if this was the goal.

Patch-set: 2