Update patch set 2
Patch Set 2:

> And why we need this? If we use a different branch in the
 > distribution it will not pull the projects artifacts created by
 > this job, right? I saw some comment from Vratko regarding the
 > ability of passing a PATCHREFSPEC to the integration-distribution-test-{stream}
 > job, but that is not this change if this was the goal.

Vratko added the change where we can specify a refspec for distribution.  if no
refspec is passed then we just clone distribution master and use that to create
the distro.  So, it will always be on master.

but, if we want to build a distro based on some stable/lithium project patches,
then we need to be able to use the stable/lithium branch of the distribution
project to build it.

you can see that here:
https://jenkins.opendaylight.org/sandbox/job/integration-multipatch-test-beryllium/10/

it's a stable/lithium OVSDB patch and stable/lithium branch of distribution to create
it.

Patch-set: 2
Label: Verified=0