Update patch set 2
Patch Set 2:

> but in that case we do not have the integration-multipatch-test-stable-lithium?

ah.  good point  in my mind I was wanting to move away from 
multiple jobs and just have one job with this branch variable.
that is more work to be done, I know.

otherwise, if we keep the multiple jobs, how do we get access
to the branch name inside the script?

Patch-set: 2