Update patch set 1
Patch Set 1: Code-Review-1

this depends on an int/test patch that I'm still working through
in the sandbox.

https://git.opendaylight.org/gerrit/#/c/29471/

I think this work is close now, so any reviews are appreciated but we
need to hold off merging this until the int/test work is done and
upstream.

Patch-set: 1
Label: Code-Review=-1