Update patch set 3
Patch Set 3: Code-Review+1

Luis, can you check how this looks to you?

I think I prefer a smaller job count to start with and since this is 
not an auto-job and only run on request I think it fits better this
way.

this sandbox job is using an ovsdb patch from stable/lithium and uses stable/lithium distribution branch:
https://jenkins.opendaylight.org/sandbox/job/integration-multipatch-test/1

this sandbox job uses master for both (master is default for distro branch):
https://jenkins.opendaylight.org/sandbox/job/integration-multipatch-test/1

Patch-set: 3