Update patch set 2
Patch Set 2:

Unfortunately this patch puts us over the time limit for when a slave has to be online. I think it's because Firefox likely pulls in a ton of other GUI related dependencies. We'll have to bake this into the Vagrant image so that we don't get hit with the time limit.

Patch-set: 2