De-revert: Set pass-if to 100% in csit-related jobs
This is a de-revert of Change 30390,
but this time with override-votes set to true.

Some jobs had 0%, but the agreement is to report
yellow dot (as opposed to blue dot)
whenever any test case failed.

Explicitly set UNSTABLE to vote verified +1
as most test suites lack ability
to ignore known non-blocking bugs.

Change-Id: I5313c4bbf283a7970f9099517484834cd20c0886
Signed-off-by: Vratko Polak <vrpolak@cisco.com>