Update patch set 1
Patch Set 1:

I've been thinking of making the sonar jobs run weekly as well but I'm not sure about that one. I feel for the most part not many people check the dailies so moving it to a weekly job might be a good idea. I don't expect any push back but might be worth starting a thread on release mailing list just to make sure.

Vratko can you start a thread for this one? We'll let it sit for a week or 2 and if nothing major comes out of it I say we turn it into a weekly job. Would be worth asking about the sonar job as well. Considering the sonar job has a trigger "run-sonar" I too feel the running daily may not be too much value.

Patch-set: 1