Jenkins administrative groovy scripts
Make slave cleanup (destruction) groovy scripts accessible to anyone
that has script console access in Jenkins. Mostly this is just me
publishing the scripts so that Thanh has access and Google picks them up
when it indexes our mirror for others that may run into similar problems
that we have when trying to cleanup slaves that are not cleaning up
properly.

Change-Id: I86c5f773879eaff25e4579445ad2b8036bb080f6
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>