Update patch set 3
Patch Set 3:

> > I think you might be getting hit with the JJB bug that we've been
 > > discussing in https://git.opendaylight.org/gerrit/29170/
 > >
 > > Can you try Vratko's suggestions from Gerrit 29170 and see if you
 > > can get it working?
 > 
 > Yes, it works! And I have successfully run one with
 > https://git.opendaylight.org/gerrit/#/c/30553/ . Thanks a lot!
 > 

That's great! Thanks for confirming. We've also seen a fix for that bug get merged in JJB this week so in the future when they release a new version either 1.3.1 or 1.4.0 we won't have to worry about this any more. There's been a request to openstack to produce a new release but I'm not sure how soon it will happen.

Patch-set: 3
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0