Update patch set 3
Patch Set 3:

My only concern is that this job might over take queues that we need for verify jobs. How often does this run and would it getting queued cause this problem?

Also is anything else still using dynamic_controller if this job stops using it?

I'm also look into the Priority Sorter Plugin which should help us with this issue in the future assuming it works. I just need some time to evaluate it.

https://wiki.jenkins-ci.org/display/JENKINS/Priority+Sorter+Plugin

Patch-set: 3