Update patch set 1
Patch Set 1: Code-Review+1

Here is an example of how it works:
https://jenkins.opendaylight.org/sandbox/job/integration-distribution-deploy-beryllium/6/console


there is a big downside.  In order to catch this problem, we have to wait for it to show up.  I've tested that 5 minutes seems to be long enough (took a little under 4 on a very large test system of mine).

For now, I've only added this work in the distribution-deploy job.  I'm not sure the best place for it.  Will this be enough?  At least 
we'll have some place to catch it.

I think it might be better to make this check for *any* output on
the karaf console, but for now it's just looking for BindException.

Having said and done all of this, we really need to figure out how
these messages are only coming to the console and not the log file
to see if we can also fix that.  I'm not sure where to start on 
that problem.

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1