Updated MVN_GOALS for VTN project.
The following patch for VTN project has changed the configuration about
how to build modules in VTN project.
https://git.opendaylight.org/gerrit/#/c/31096/
In order to build all modules in VTN project, we need to activate
vtncoordinator and vtnapps profiles. Therefore, this patch adds the two
profiles into the list of profiles to be activated.

Change-Id: I78d073befd87d51fc3ca73f7ac075c01ccc261f9
Signed-off-by: Hideyuki Tai <Hideyuki.Tai@necam.com>