Update patch set 1
Patch Set 1:

The only purpose for passing -Dstream={branch} is so that Maven sites are pushed to the correct directory in Nexus sites. For example odlparent's the only project who uses this so far (I've been slow at deploying this to other projects).

 https://nexus.opendaylight.org/content/sites/site/org.opendaylight.odlparent/

The flag determines which directory to publish the site artifacts to. Currently beryllium or pre-boron.

If this flag is not passed it will create a directory called "latest" by default. So really the only job that truely needs this is Merge jobs. I do it in Verify jobs just so that it's verifying by creating the same directory that a merge job would have.

Patch-set: 1