Update patch set 1
Patch Set 1:

Ok so in the autorelease job the provide configuration files builder step appears as a "post step" which is after the maven command has run. I think this is why it's not picking it up. This patch put it in the prebuilders section though so I'm confused why JJB is not placing it in the right spot. Going to poke at this some more...

Patch-set: 1