Update patch set 1
Patch Set 1:

Disabled is correct. The downside to this flag though is once it's set it always has to be set. If you remove the option JJB won't detect it as a change in config and keep the previous setting.

I'd prefer to remove the job altogether if it's not needed.

Patch-set: 1
Label: Code-Review=0