Update patch set 2
Patch Set 2:

I'm stumped on this one. It is only reproducible when using Jenkins to run the test command. Logging into the master silo via ssh and running the same commands do not reproduce the issue.

I've decided to revert to 1.3.0 for the releng silo for now. Maybe we can update once we move into the private cloud and the issue will go away when we switch to CentOS 7 for the Jenkins server.

Patch-set: 2