Update patch set 1
Patch Set 1:

I tried to keep this short and sweet. If we can agree on the minimal set of info we need to in order for the user to be routed to the correct pages then I can work on pushing this template to all ODL projects.

Patch-set: 1
Label: Verified=0