Update patch set 2
Patch Set 2: Code-Review+2

Yeah, "md" is hardcoded in files such as [0] and it looks like "markdown" extension would need a separate *Module class (and perhaps few naming conflicts would need to be resolved).

I have created a Jira issue for Doxia [1], but we have to stick with ".md" for now.

[0] http://www.javased.com/index.php?source_dir=maven-doxia/doxia-modules/doxia-module-markdown/src/main/java/org/apache/maven/doxia/module/markdown/MarkdownSiteModule.java
[1] https://issues.apache.org/jira/browse/DOXIA-535

Patch-set: 2
Label: Code-Review=+2