Make the ODL download more verbose
This changes the ODL downloading commands to ask them to
generate progress reports.

One of the informations that is provided is the total time
taken to download the ODL which turns to be quite useful
when investigating problems stemming from slow download
times.

The second useful effect is that it makes the places
in the console output where ODL is being downloaded far
easier to quickly locate (hold PgUp/PgDn until you see
the characteristic progressbar output, then release
PgUp/PgDn and you are roughly at the place where one
of the ODL downloads was happening).

Change-Id: Iee8eff30aeab28dd8c98c15ea909dc67a7ee8ef2
Signed-off-by: Jozef Behran <jbehran@cisco.com>