Update patch set 2
Patch Set 2:

> Why would .md be better for GitHub

When I see two (even slightly) different formats, I like to assign different extensions for them. In this case, one format requires backslash before underscore (to avoid italics), the other does not require it (but maybe it allows the backslash, I have not tested that).
If there is a way to write a file to make it look the same on standard and GitHub-style markdown implementations, we can use that way (and stop caring about extensions too much).

I am not sure which of the two formats are implemented by Doxia plugin (not relevant for this Change).

Patch-set: 2