Update patch set 4
Patch Set 4: Code-Review-1

> I've decided to revert to 1.3.0 for the releng silo for now.

This is why I do not recommend merging this right now. We do not want a situation when a job which works on Sandbox but not on the releng silo.

Unless the following is actually fixed by this change:
> the runaway recursion in [0] if fault of bad .yaml file or JJB itself.

Ideally, builder-merge virtualenv should be changed to use exactly what current requirements.txt say, check the runaway issue again, and think about upgrade to 1.4.0 after that.

By the way, aside on the following warning, 1.4.0 works for me on Sandbox.

> DeprecationWarning: get_plugins_info() is deprecated, use get_plugins()

[0] https://jenkins.opendaylight.org/releng/job/builder-verify-jjb/1926/console

Patch-set: 4
Label: Code-Review=-1