Update patch set 4
Patch Set 4: Code-Review+1

> > I've decided to revert to 1.3.0 for the releng silo for now.
 > 
 > This is why I do not recommend merging this right now. We do not
 > want a situation when a job which works on Sandbox but not on the
 > releng silo.

So I figured out the cause finally. It's due to overriding the PYTHONPATH variable that we did long ago before we started using virtualenvs for everything. The overrided PYTHONPATH points to a python2.6 version of Python which JJB recently removed support for. This was a global setting on Jenkins master which explains why I couldn't reproduce the issue anywhere except inside a Jenkins job.

Releng master is now running 1.4.0 as we expect.

I cannot remember which project we hacked the PYTHONPATH for but they should have migrated to using virtualenv to get their dependencies now. So the plan is to wait and see if any projects report that their python jobs are failing now and fix their job to use virtualenv. Hopefully there won't be any.

Patch-set: 4