Update patch set 4
Patch Set 4:

Ok, I spent some time tonight with git bisect and was able to pin-point the commit that introduced the issue. It's [1]. Still doesn't explain why it's only reproducible through Jenkins and not through an ssh session to the same machine but I've also confirmed it's reproducible on Python 2.7 through Jenkins. If I used Python 3.3 it didn't run into this issue. I'm going to look into getting Python 3.3 setup in the Jenkins environment.


[1] https://review.openstack.org/#/c/183939/

Patch-set: 4