Update patch set 2
Patch Set 2: Code-Review+1

> set an echo in between

Yes, output of this change may be confusing, but it is still a step forward.

More pleasant solution would be to first check karaf.log size. If it is 1M or smaller, keep whole. If it is larger, create two files karaf_head.log and karaf_tail.log
But I believe there are more pressing things to do than such cosmetic improvements.

Patch-set: 2