Update patch set 2
Patch Set 2: Code-Review+1

This may need to be voted on by affected projects, my +1 is for integration/distribution and integration/test.

I have a feeling that some of MVN_GOALS and MVN_OPTS lines are no longer needed, or they can be used globally; but I have included enough edits to this Change already.

Patch-set: 2