Update patch set 2
Patch Set 2: Code-Review-1

Oh, I vaguely remember now. Even when Helium Bugs are currently being marked as WONTFIX as Helium is no longer supported, it is still supported with respect to security bugs.
If a security bug appears, it will be nice to have all jobs (and Nexus artifacts) ready for processing the fix, and merge+csit runs over weekend are not that taxing.

Patch-set: 2
Label: Code-Review=-1