Update patch set 2
Patch Set 2:

> I have created this Change by accident, I realized too late that I
 > am no longer fixing jobs only for integration/distribution. :)
 > 
 > Jamo has a task for this [0], so I will not continue on it.
 > 
 > TODOs:
 > * Verify on Sandbox this way actually works.
 > * Perhaps "jre" is better name to distinguish from "jdk"
 > and allow "build+test" style jobs to chose non-matching versions.
 > * Separate secondary edits to standalone Changes.
 > * Perhaps each bump to openjdk8 has to get +1 from affected
 > project?
 > * After merges, make sure to notify open changes to expand their
 > stream structure.
 > 
 > [0] https://trello.com/c/iR52y3zH/250-java-8-for-csit-jobs


This looks like it's changing all CSIT jobs to specifically use jdk8
for Boron and jdk7 for everything else.  That might be the right 
approach.  Are we sure we don't want something like this?

Also, Chaudhry (now on this review) is doing the work to make the 
distribution-test job (which triggers all these CSIT jobs) use the
*other* version.  So, it could use jdk7 for Lithium and Beryllium
on a daily basis just for the coverage.  I'm assuming we don't want
anything jdk7 for Boron+.

As for the TODOs: the build+test is a good idea, but we are going
to be getting it for free right?  meaning, we are having Li/Be built
with 7, but we'll end up having it run with 8.

Patch-set: 2
Label: Code-Review=0