Update patch set 2
Patch Set 2: Code-Review-1

Vratko: Can you please answer the following questions to address my concerns/confusions. 

Isn't the variable name should be 'jdkversion' instead of 'jdk'? As the macro which I added in the previous patch to the below file was 'jdkversion'.

https://github.com/opendaylight/releng-builder/blob/master/jjb/integration/integration-macros.yaml#21

Secondly! You did this effort to make Java Version configurable. Why we have to add the parameter in each of the job? Aren't we already defaulted these jobs to use openjdk7 in the below file:

https://github.com/opendaylight/releng-builder/blob/master/jjb/integration/integration-templates.yaml

We can simply pass a parameter from the master to the calling jobs in case of boron. I did this in the following job
https://jenkins.opendaylight.org/sandbox/job/integration-distribution-test-beryllium/1/
and it makes other jobs jdk version set to openjdk8.

Patch-set: 2