Add logic to skip closing the Nexus staging repo
In order to sign artifacts we need to skip closing the staging repo
after it is produced. This allows the gpg key signatures to be uploaded
into the staging repo.

This patch uses the existing DATESTAMP variable to determine if the
staging repo should be created open or closed. If a DATESTAMP is
true then close the staging repo, otherwise it is likely a release
candidate so skip closing the staging repo.

Change-Id: I2330c467824c0b50d70923aedbe20e56ac9cb0e9
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>