Do not leave staging repo open
Nexus does not publish a staging repo if the repo is left in the Open
state. Unfortunately this means we have to decide on what's more
important.

1) Automated tests can trigger immediately after autorelease produces
   new build.
2) Ability to add additional artifacts to a Maven repo (such as GPG
   signatures)

Case 1 can only work if we Close the Staging repo. Case 2 requires us to
leave the Staging repo in Open state so that we can push gpg signatures
up after the build completes.

The workaround for Case 1 is the manually launch the
integration-distribution test once Helpdesk manually makes the staging
repo available post build. Which makes something that used to be
automated to now be a manual task which isn't ideal either.

Change-Id: I00c210cec159c83e7d3badf16306586541e364dc
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>