Allow staging repos to remain open again
This reverts commit d923faff8b1fc8a645f31076de013326b3117e00.

Found a solution that hopefully works allowing us to continue with the
automated integration job and allows us to keep the staging repo open
for GPG signing.

We're hosting the staging repos through a web backend to allow us to
access the staged repos while at the same time allow the repos to be in
Open state in Nexus.

This patch updates the download path for the artifacts needed by the
integration distribution job.

Change-Id: Ic1700d1cc8ffe0e89f94c56f4da0c7a6be84bea2
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>