Remove mvn31 verify jobs in Beryllium
Working with the docs project we would really like to start pushing
javadocs to Nexus, especially the apidocs. As we found in Gerrit 27711
one of the issues we're having is with generating maven-sites against
Maven 3.1. In this version of Maven, site generation fails with:

[WARNING] The POM for org.apache.felix:maven-bundle-plugin:jar:3.0.0 is
missing, no dependency information available

We'd like to propose that we don't test with Maven 3.1.x series in
Beryllium so that we can cherry-pick the patch from Gerrit 27711 in
Beryllium in hopes that in the next Beryllium release (SR1) we'll be
able to provide Java apidocs for all projects in Nexus.

See:
https://git.opendaylight.org/gerrit/27711/

Change-Id: I7b725ab270f3cbfdb4b1204329824b2b910fdf62
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>