Update patch set 7
Patch Set 7:

I was waiting for a decision on if the job should be default for every project or not. I guess it can't help to just make the change now. It's now set to default for every project and a project will need to add the JOB_TEMPLATES parameter to disable it. Let me know if it should be the other way around.

Patch-set: 7
Label: Verified=0