Update patch set 12
Patch Set 12:

That ERROR is fine. We don't maintain the settings files on the sandbox since they would all be empty anyway as we do not want to give the Sandbox the privileges to publish to Nexus.

Patch-set: 12
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0