Update patch set 12
Patch Set 12: Code-Review+2

There was an e-mail [0].

> We've prepared a patch [1] that we would like to merge tomorrow (Tuesday) if there are no strong objections.

Giving +2, but not merging as it is not clear to me which part of today is the deadline.

> This job takes about 15 minutes to run so should typically finish sooner than your project verify job.

I see 159 jobs in the queue right now, but all of them are csit jobs, there are multiple slaves with idle executors, both -build- and -verify-.
But this improves quality of verification, which is worth it even in busy periods where this will cause additional delays.

[0] https://lists.opendaylight.org/pipermail/release/2016-February/005424.html

Patch-set: 12
Reviewer: Gerrit User 2046 <2046@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+2