Update patch set 13
Patch Set 13:

> karaf probably used 7 from JAVA_HOME, but can we know for sure?

We know how it is parsed in bin/karaf [0], it is not straightforward.

I have started several runs on Sandbox. All red dots, at least because sshpass was not located, which is a Sandbox bug as jenkins_scripts install it, both in controller.sh and builder.sh

Beryllium build works on Java 8 [1], but Boron build gets stuck at 404 on Java 7 [2].

[0] https://git.opendaylight.org/gerrit/gitweb?p=controller.git;a=blob;f=karaf/opendaylight-karaf-resources/src/main/resources/bin/karaf;h=232325ea99f66884d145ef2ca97a9d3f95991085;hb=refs/heads/stable/beryllium
[1] https://jenkins.opendaylight.org/sandbox/job/integration-distribution-deploy-beryllium/2/console
[2] https://jenkins.opendaylight.org/sandbox/job/integration-distribution-deploy-boron/1/console

Patch-set: 13