Update patch set 2
Abandoned

For now "pip freeze" is in include-raw-integration-install-robotframework.sh and there also is a suite that does all sort of queries on test environment (stuff like "how big is the disk") but that suite is nowhere referenced (suggestion from Vratko: Add to some *-verify-* job so it gets executed on Releng as well). Stashing for now, when something shows up that the two above items can't cover, may be resurrected.

Patch-set: 2
Status: abandoned