Add job for the bug 5383 checking
The bug 5383 needs "almost empty Karaf" so it cannot run in
another job that configures its featues in the
"install-features" setting. To be exact, "odl-restconf" must
be present and "netconf-connector" is present because that
does not prevent the suite from catching the bug and the bug
will be caught faster.

Change-Id: I9ad1c011735efc020db15a63dab1ba11bde06401
Signed-off-by: Jozef Behran <jbehran@cisco.com>