Use Maven's deploy command to stage artifacts
The current autorelease staging script is problematic because it pulls
artifacts from your local maven repo which can contain artifacts that
weren't built by the build process (yangtools pulls old artifacts for
testing).

This is also problematic because this means we're ignoring the artifacts
that explicitly state that they should be skipped during the deploy
phase.

Use Maven's deploy goal to push to an alternative staging URL which can
then be used to push a staging repo containing only artifacts built
during the build.

Change-Id: Icb976a72929569dec5fbdfe6fe7bdc8880ff4284
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>