Update patch set 1
Patch Set 1: Code-Review+1

Let Thanh +2 and merge this.  Not sure if the c6 VM is still available
for us or not.  If so, merge it, and we'll figure out how to get
cbench working on ubuntu VM.

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1