Update patch set 2
Patch Set 2: Code-Review+1

I'm ok with this being merged.  I wanted to follow through and at least get to the same point, as technically projects are supposed to be responsible for this.  I've done that, so I'm satisfied.  Note that the variance in generated cfg file indicates that some work on the jjb-init-project script is required.

Patch-set: 2
Reviewer: Gerrit User 3576 <3576@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1